PLAT-376: do not compile flacky test on windows
